waffle-wisdom/python_proj/requirements.txt
openai==1.66.2
functions-framework==3.8.2
google-cloud-secret-manager==2.23.2